syncFocusToMacOs only when AeroSpace tries to change the focus
This commit simplifies the logic on how the focus gets returned back to
the macOS

Motivation is misbehaved completion popup in DBeaver
https://github.com/nikitabobko/AeroSpace/issues/106

DBeaver creates a regular window for it's completion popup instead of a
real popup window and doesn't focus the fake popup window => since it's
a new real window, AeroSpace thinks that it should be in focus
(automatic propogation to the mru in TreeNode) => by the end of the
refreshSession we notice that there is a mismatch between nativeFocused
and focusedWindow ("AeroSpace perspective"), so AeroSpace commands to
focus the focusedWindow to resolve the mismatch

Maybe it wasn't the best idea trying to resolve nativeFocused vs
"AeroSpace perspective" mismatches. Just let the mismatch exist. Apps
can behave in a variety of ways. The apps might create fake windows or
real popups (that's what `nativeFocused != nil` hack was here for), keep
or change the focus, your never know. So it's better not to touch the
focus in case of a mismatch, and just let happen what was supposed to
happen.

We still:
1. Propogate the focus back to the macOS when focused window is
   explicitly changed in AeroSpace model
2. Sync focus from macOS to AeroSpace. When the focus is changed
   externaly (e.g. cmd+tab), we change AeroSpace mental model

It looks like it should be enough, no need to do more than that

An alternative fix is to drop automatic mru propogation when attaching
new nodes in TreeNode. The automatic propogation causes issues in
FocusCommand as well when we want to temporarily attach floating
windows.

But I'm still not sure about the correctnes of the alternative fix