Remove launchctl calls in start on login routine
https://github.com/nikitabobko/AeroSpace/issues/479
closes https://github.com/nikitabobko/AeroSpace/pull/480

As described in #479, the calls to launchctl are broken due to invalid
file paths, but fixing them introduces a worse bug.

As macos/launchd will automatically load a plist if it's present in the
launch agents directory on load, it's enough that we simply write/delete
the file as needed. For now at least, as there is no regression in
functionality.