Implement focus dfs-next/dfs-prev.
This allows switching to the next/prev window from the current one in
the depth-first order of the windows in the current workspace tree.

This is convenient, as it allows users to shift focus through a
workspace's windows in a more predictable way than the directional
movement commands, where the move target can depend on invisible
most-recently-used state.

_fixes https://github.com/nikitabobko/AeroSpace/issues/248