Add max width settings for single-window workspaces and fullscreen
## The problem

When using an ultrawide monitor, a single window in a workspace can
often be too wide.

## The solution

This commit introduces two new settings and an additional flag to the
fullscreen command that make single windows behave better on wider
monitors. The changes are fully backwards compatible.

New settings:

1. `single-window-max-width-percent`: Sets maximum width percentage
	 (1-100) for windows which are the only window in a workspace, and
	 also limits the width of the workspace when in fullscreen mode
2. `single-window-exclude-app-ids`: An array of app bundle IDs to
	 exclude from the width limit

New flag to the `fullscreen` command:

- `--no-max-width`: Ignores single-window width constraints when in
	fullscreen mode

## Additional notes

With these settings and the flag, the user can separately bind
fullscreen in both the normal (constrained) mode and in the
unconstrained mode, allowing quick switching between the two. (This is
how I use it and it feels very natural.)

This addresses #60.