Add `--stdin` flag and require it when providing a list to the `workspace` command via stdin
Issue: https://github.com/nikitabobko/AeroSpace/issues/1683

Added the `--stdin` flag to the `workspace (prev|next)` command. When
using this command, the flag is now required if the user wants to input
the list of workspaces from stdin. If there is input from stdin but the
flag is not specified, the command will display a warning.