Add `--stdin` flag and require it when providing a list to the `workspace`/`move-node-to-workspace` command via stdin
closes https://github.com/nikitabobko/AeroSpace/pull/1684

Issue: https://github.com/nikitabobko/AeroSpace/issues/1683

Added the `--stdin` flag to the `workspace (prev|next)` command. When
using this command, the flag is now required if the user wants to input
the list of workspaces from stdin. If there is input from stdin but the
flag is not specified, the command will display a warning.

Added the --no-stdin flag to the workspace (prev|next) command. When
using this command, the flag is now required if the user wants to ignore
input from stdin when such input exists. If there is input from stdin
but the flag or `--stdin` flag is not specified, the command will
display a warning.