Add `--no-stdin` flag and require it when stdin input exists without `--no-stdin` or `--stdin` flag in `workspace (prev|next)` command
Issue: https://github.com/nikitabobko/AeroSpace/issues/1683

Added the --no-stdin flag to the workspace (prev|next) command. When
using this command, the flag is now required if the user wants to ignore
input from stdin when such input exists. If there is input from stdin
but the flag or `--stdin` flag is not specified, the command will
display a warning.