Implement automatic key-mapping preset based on currently selected keyboard layout
By default, all key bindings in the AeroSpace config get interpreted based on a qwerty layout. When switching keyboard layouts, this means that AeroSpace's keyboard shortcuts don't adjust based on the new location of the key as defined by the keyboard layout.

This usually isn't a big problem, as `key-mapping.key-notation-to-key-code` can be used to adjust the layout. However, when working with a custom keyboard that is remapped on the firmware level (i.e. anything running ZMK), in conjuncion with a keyboard that is remapped in software (i.e. builtin MacBook keyboard), this results in significant issues.

For example, let's consider someone using the Dvorak layout with the `dvorak` preset configured in AeroSpace. Let's assume they have some command bound to `alt-o`. Due to the `key-notation-to-key-code` mapping, AeroSpace starts listening for events concerning the key code 1 (which on qwerty corresponds to 's').

When they press `alt-o` on the builtin MacBook keyboard, they press the physical `s` key, resulting in the key code 1.
However, when they press `alt-o` on their other keyboard that is remapped to Dvorak in the keyboard firmware, the keyboard sends key code 31 (which on qwerty corresponds to 'o') instead. Thus AeroSpace won't run the command.