default corner option
I have a configuration with three monitors like this:

(two horizontal stacked and a big vertical on the right)

|   |   |
|---|   |
|   |   |

With this configuration the default behavior is really not ideal.
Whereas if we just add a default corner configuration we can have
something that works as expected.

This pull-request is just adding